back to js webpack bc esm incompatibility
