+++ demo/index.html
@@ -2,7 +2,7 @@
   <head>
     <title>React Modal</title>
     <link rel="icon" type="image/png" href="https://facelessui.com/favicon.png" />
-  </head>
+  <script defer src="demo.bundle.js"></script><script defer src="demo.bundle.js"></script><script defer src="demo.bundle.js"></script></head>
   <body>
     <noscript>Enable JavaScript</noscript>
     <div id="root"></div>
