+++ package.json
@@ -44,10 +44,11 @@
     "ux"
   ],
   "scripts": {
-    "build": "yarn lint && tsc --project tsconfig.build.json",
-    "build:demo": "yarn lint && webpack --config webpack.build-demo.config.ts",
-    "dev": "webpack-dev-server --hot --config webpack.dev.config.ts",
-    "lint": "eslint .",
+    "build": "yarn lint:src && tsc --project tsconfig.build.json",
+    "build:demo": "webpack --config webpack.build-demo.config.js",
+    "dev": "webpack serve --config webpack.dev.config.js",
+    "lint:src": "eslint ./src",
+    "lint:demo": "eslint ./demo",
     "test": "jest"
   },
   "peerDependencies": {
